server/bridge/requirements.txt
paho-mqtt>=1.6.0
influxdb>=5.3.0
cbor2>=5.4.0
